% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPM_withcomps.R
\name{ASPM_withcomps}
\alias{ASPM_withcomps}
\title{Make ASPM runs}
\usage{
ASPM_withcomps(
  Path,
  ASPM_Path,
  Rdevs,
  comp_fleet,
  Hessian = FALSE,
  dat_name = "BET-EPO.dat",
  ctl_name = "BET-EPO.ctl",
  ss_name = "ss.exe",
  par_line = 6
)
}
\description{
\code{ASPM_withcomps} This function generates ASPM
}
