% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_indexcomps.R
\name{CCA_indexcomps}
\alias{CCA_indexcomps}
\title{Make CCA runs}
\usage{
CCA_indexcomps(
  Path,
  CCA_Path,
  Index_fleet,
  Hessian = FALSE,
  dat_name = "BET-EPO.dat",
  ctl_name = "BET-EPO.ctl",
  ss_name = "ss.exe",
  par_line = 6
)
}
\description{
\code{CCA_indexcomps} This function generates CCA
}
