% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Check whether the SS model is converged without parameters hitting the bound}
\usage{
check_convergence(Path)
}
\description{
\code{check_convergence} This function checks whether the SS model is converged without parameters hitting the bound
}
