% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareEffN.R
\name{CompareEffN}
\alias{CompareEffN}
\title{Plot Pearson residuals for SS runs}
\usage{
CompareEffN(Path, myFleet, model_name)
}
\description{
\code{CompareEffN} This function compare estimated effective sample size in a table
}
