% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareSelex.R
\name{CompareSelex}
\alias{CompareSelex}
\title{Compare size selectivity for SS runs}
\usage{
CompareSelex(Path, FleetNums, Legend)
}
\description{
/code{CompareSelex} This function plot compare selectivity curves among models
}
