% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_compare.R
\name{cpue_compare}
\alias{cpue_compare}
\title{Compare cpue from different VAST models}
\usage{
cpue_compare(
  Path,
  Legend,
  Region,
  Save_Path,
  rescale = TRUE,
  ylabel = "CPUE",
  xlabel = "",
  ylim,
  xlim,
  CV = FALSE,
  w = 15,
  h = 8,
  smooth = FALSE,
  version = "New",
  s = 0.1,
  Figure_name = "CPUE_Compare",
  Year_first = 0,
  Year_last = 2500,
  Legend_name = "Index"
)
}
\description{
/code{cpue_compare} This function plot the comaprison of standardized cpue from different VAST models
}
