% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_cv_compare.R
\name{cpue_cv_compare}
\alias{cpue_cv_compare}
\title{Compare cpue from different VAST models}
\usage{
cpue_cv_compare(
  Path,
  Legend,
  Region,
  Save_Path,
  ylabel = "CV",
  xlabel = "",
  ylim,
  xlim,
  w = 15,
  h = 8,
  Figure_name = "CPUE_CV_Compare",
  Legend_name = "CV"
)
}
\description{
/code{cpue_cv_compare} This function plot the comaprison of standardized cpue from different VAST models
}
