% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faa_a.R
\docType{data}
\name{Dir}
\alias{Dir}
\title{Fishing mortality by age group}
\format{
An object of class \code{character} of length 13.
}
\usage{
Dir
}
\description{
\code{faa_a} This function computes the fishing mortality by age group for the stock assessment report
}
\keyword{datasets}
