% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_catch.R
\name{equilibrium_catch}
\alias{equilibrium_catch}
\title{Extract equilibrium catch}
\usage{
equilibrium_catch(SS_Dir, Fishery)
}
\description{
\code{equilibrium_catch} This function extracts the expected equilibrium catch
}
