% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_NLL_npars_gradient.R
\name{extract_NLL_npars_gradient}
\alias{extract_NLL_npars_gradient}
\title{Make ASPM runs}
\usage{
extract_NLL_npars_gradient(file_path)
}
\description{
\code{extract_NLL_npars_gradient} This function generates ASPM
}
