% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jitter_results.R
\name{get_jitter_results}
\alias{get_jitter_results}
\title{Generate area code for longline}
\usage{
get_jitter_results(Dir, njitters)
}
\description{
\code{get_jitter_results} This function generates an area code for longline catch allocation
}
