% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_plot.R
\name{impact_plot}
\alias{impact_plot}
\title{Fishing impact plot}
\usage{
impact_plot(Dir, n_year, BaseName = "Base", n_fishery, title, proj_year = 0)
}
\arguments{
\item{Dir}{The directory in which the base model is stored as a subfolder (not the directory of the base model folder)}

\item{n_year}{The number of model time steps}

\item{BaseName}{The name of the base model folder}

\item{n_fishery}{The number of fishery fleets}

\item{title}{The title of the impact plot}
}
\description{
\code{impact_plot} This function automatically generates the fishery impact plot for SS v3.30
}
