% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_plot_OBJ.R
\name{impact_plot_OBJ}
\alias{impact_plot_OBJ}
\title{Fishing impact plot}
\usage{
impact_plot_OBJ(Dir, n_year, BaseName, n_fishery, title)
}
\description{
\code{impact_plot_OBJ} This function automayically generates the fishery impact plot for SS v3.23
}
