% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLcatch.R
\name{ll_catch}
\alias{ll_catch}
\title{Longline catch allocation}
\usage{
ll_catch(Grid_Catch, FSR_Catch, Species, last_year, dir)
}
\description{
\code{ll_catch} This code processes the raw LL catch data into the format for Stock Assessment
}
