% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_catch_all.R
\name{ll_catch_all}
\alias{ll_catch_all}
\title{Longline catch allocation}
\usage{
ll_catch_all(Grid_Catch, FSR_Catch, Species, last_year, dir)
}
\description{
\code{ll_catch_all} This function processes the raw LL catch data into the format for Stock Assessment
}
