% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_catch_KOR.R
\name{ll_catch_KOR}
\alias{ll_catch_KOR}
\title{Longline catch allocation}
\usage{
ll_catch_KOR(Grid_Catch, FSR_Catch, Species, last_year, dir)
}
\description{
\code{ll_catch_KOR} This function processes the raw LL catch data into the format for Stock Assessment (KOR is considered a seperate fleet)
}
