% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_fisheries_lf_joint.R
\name{ll_fisheries_lf_joint}
\alias{ll_fisheries_lf_joint}
\title{Longline fisheries' length frequency}
\usage{
ll_fisheries_lf_joint(
  JPN_size,
  KOR_size,
  Add_KOR = FALSE,
  catch_weighted = FALSE,
  Grid_Catch,
  Species,
  last_year,
  dir,
  minNsamp = 1,
  period,
  legend
)
}
\description{
\code{ll_fisheries_lf_joint} This function processes the raw LL fisheries' length frequency data into the format for Stock Assessment
}
