% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_fisheries_lf_new.R
\name{ll_fisheries_lf_new}
\alias{ll_fisheries_lf_new}
\title{Longline fisheries' length frequency}
\usage{
ll_fisheries_lf_new(
  JPN_size,
  Grid_Catch,
  Species,
  last_year,
  dir,
  minNsamp = 1,
  period,
  legend
)
}
\description{
\code{ll_fisheries_lf_new} This function processes the raw LL fisheries' length frequency data into the format for Stock Assessment
}
