% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_fisheries_lf_separate.R
\name{ll_fisheries_lf_separate}
\alias{ll_fisheries_lf_separate}
\title{Longline fisheries' length frequency}
\usage{
ll_fisheries_lf_separate(
  JPN_size,
  KOR_size,
  Grid_Catch,
  Species,
  last_year,
  dir,
  period,
  legend,
  minNsamp = 0,
  LF_name = "LF",
  plot_map = FALSE,
  rounding = "low"
)
}
\description{
\code{ll_fisheries_lf_separate} This function processes the raw LL fisheries' length frequency data into the format for Stock Assessment
}
