% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLcatch.R
\name{LLcatch}
\alias{LLcatch}
\title{Longline catch allocation}
\usage{
LLcatch(Grid_Catch, FSR_Catch, Species, last_year, dir)
}
\description{
\code{LLcatch} Allocate Longline catch by area and year-quarter for the stock assessment
}
