% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kobetable.R
\name{make_kobetable}
\alias{make_kobetable}
\title{Make Kobe table}
\usage{
make_kobetable(year1, BasePath, KobePath)
}
\description{
\code{make_kobetable} This function automayically generates the table for kobe plot
}
