% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeManagTable.new.R
\name{makeManagTable.new}
\alias{makeManagTable.new}
\title{Make the management table based on SS output}
\usage{
makeManagTable.new(Path, FFleets, FstdPath, FlimitPath)
}
\description{
\code{makeManagTable.new} This code make the mangament table for IATTC stock assessments
}
