% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeManagTable.R
\name{makeManagTable}
\alias{makeManagTable}
\title{Make the management table based on SS output}
\usage{
makeManagTable(Path, FFleets, dynamicS0, Dynamic_Path)
}
\description{
\code{makeManagTable} This code make the mangament table for IATTC stock assessments
}
