% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeManagTable_SAC11.R
\name{makeManagTable_SAC11}
\alias{makeManagTable_SAC11}
\title{Make the management table based on SS output}
\usage{
makeManagTable_SAC11(Path, FFleets, stdPath, dynPath)
}
\description{
\code{makeManagTable_SAC11} This code make the mangament table for SAC 11; modified based on makeManagTable.new
}
