% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareSelex.R
\docType{data}
\name{Path}
\alias{Path}
\title{Compare size selectivity for SS runs}
\format{
An object of class \code{character} of length 2.
}
\usage{
Path
}
\description{
/code{CompareSelex} This function plot compare selectivity curves among models
}
\keyword{datasets}
