% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_cdf.R
\name{pdf_cdf}
\alias{pdf_cdf}
\title{Fishing mortality by age group}
\usage{
pdf_cdf(x, DF)
}
\description{
\code{pdf_cdf} This function computes the fishing mortality by age group for the stock assessment report
}
