% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot fisheries catch}
\usage{
plot_catch(Path, Save_Path, fisheries, label)
}
\description{
/code{plot_catch} This function automayically plot catch time series by fishery
}
