% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dSBR.R
\name{plot_dSBR}
\alias{plot_dSBR}
\title{Plot quarterly SBR}
\usage{
plot_dSBR(
  SS_Dir,
  lyear,
  fyear,
  legend,
  Save_Dir,
  ymax,
  figure_name = "",
  title = "",
  xlim,
  alpha = 0.1,
  ylabel = "",
  xlabel = ""
)
}
\description{
\code{plot_dSBR} This function plots quarterly SBR for the stock assessment report
}
