% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_R.R
\name{plot_R}
\alias{plot_R}
\title{Plot quarterly and annual recruitment}
\usage{
plot_R(SS_Dir, lyear, fyear, legend, Save_Dir)
}
\description{
\code{plot_R} This function plot quarterly and annual recruitment for the stock assessment report
}
