% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_R0profile.R
\name{plot_R0profile}
\alias{plot_R0profile}
\title{Plot the R0 likelihood profile for the bigeye model}
\usage{
plot_R0profile(Path, R0, R0_MLE, Fleet_comps)
}
\description{
\code{plot_R0profile} This function plots the R0 likelihood profile for the bigeye model
}
