% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_R0profile_YFT_EPO.R
\name{plot_R0profile_YFT_EPO}
\alias{plot_R0profile_YFT_EPO}
\title{Plot the R0 likelihood profile for the bigeye model}
\usage{
plot_R0profile_YFT_EPO(Path, R0, R0_MLE, Fleet_comps = 1:37)
}
\description{
\code{plot_R0profile_YFT_EPO} This function plots the R0 likelihood profile for the bigeye model
}
