% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_R0profile_YFT_NE.R
\name{plot_R0profile_YFT_NE}
\alias{plot_R0profile_YFT_NE}
\title{Plot the R0 likelihood profile for the bigeye model}
\usage{
plot_R0profile_YFT_NE(Path, R0, R0_MLE)
}
\description{
\code{plot_R0profile_YFT_NE} This function plots the R0 likelihood profile for the bigeye model
}
