% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_R_PO.R
\name{plot_R_PO}
\alias{plot_R_PO}
\title{Plot quarterly and annual recruitment}
\usage{
plot_R_PO(SS_Dir, legend, Save_Dir, area, figure_name)
}
\description{
\code{plot_R_Po} This function plot quarterly and annual recruitment for the stock assessment report
}
