% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Rdev.R
\name{plot_Rdev}
\alias{plot_Rdev}
\title{Plot quarterly and annual recruitment}
\usage{
plot_Rdev(SS_Dir, lyear, fyear, legend, Save_Dir, ymax, figure_name, xlim)
}
\description{
\code{plot_Rdev} This function plot quarterly and annual recruitment for the stock assessment report
}
