% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retro.R
\name{plot_retro}
\alias{plot_retro}
\title{Plot retrospective pattern}
\usage{
plot_retro(SS_Dir, lyear, fyear, Save_Dir)
}
\description{
\code{plot_retro} This function plot retrospective SB and R for the stock assessment report
}
