% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SB.R
\name{plot_SB}
\alias{plot_SB}
\title{Plot quarterly SB}
\usage{
plot_SB(
  SS_Dir,
  lyear,
  fyear,
  legend,
  Save_Dir,
  ymax,
  figure_name = "",
  title = "",
  xlim,
  alpha = 0.1
)
}
\description{
\code{plot_SB} This function plots quarterly SB for the stock assessment report
}
