% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SBR.R
\name{plot_SBR}
\alias{plot_SBR}
\title{Plot quarterly SBR}
\usage{
plot_SBR(
  SS_Dir,
  lyear,
  fyear,
  legend,
  Save_Dir,
  ymax,
  figure_name = "",
  title = "",
  xlim,
  alpha = 0.1,
  ylabel = "",
  xlabel = ""
)
}
\description{
\code{plot_SBR} This function plots quarterly SBR for the stock assessment report
}
