% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCAAL.R
\name{PlotCAAL}
\alias{PlotCAAL}
\title{Plot conditional age at length}
\usage{
PlotCAAL(Path = Path, AgeBins = AgeBins)
}
\description{
\code{PlotCAAL} This function plots conditional age at length for SS runs
}
