% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotComps.R
\name{PlotComps}
\alias{PlotComps}
\title{Plot fisheries comps}
\usage{
PlotComps(Path, Save_Path, Fleets, size, label, width = 10,
  height = 10)
}
\description{
/code{PlotFComps} This function plot fisheries aggergated comps
}
