% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCompsFit.R
\name{PlotCompsFit}
\alias{PlotCompsFit}
\title{Plot the residual of mean length for each fleet}
\usage{
PlotCompsFit(Path, fleets_len, fleets_size, span)
}
\description{
\code{PlotCompsFit} This function plots the residual of mean length for each fleet
}
