% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSelex.R
\name{PlotSelex}
\alias{PlotSelex}
\title{Plot size selectivity for SS runs}
\usage{
PlotSelex(Path, replist, FleetNums, numRows, numCols, w, h)
}
\description{
\code{PlotSelex} This function plot estimated selectivity curves for specified fleets
}
