% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpline.R
\name{PlotSpline}
\alias{PlotSpline}
\title{Plot size selectivity for SS runs}
\usage{
PlotSpline(Path, Save_Path, Fleet, model_name, fyear, lyear, dim, w, h)
}
\description{
\code{PlotSpline} This function comapres estimated and emperical selectivity curves for specified Fleet
}
