% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpline_PO.R
\name{PlotSpline_PO}
\alias{PlotSpline_PO}
\title{Plot size selectivity for SS runs}
\usage{
PlotSpline_PO(Path, Save_Path, Fleet, model_name, f, fyear, lyear, dim, w, h)
}
\description{
\code{PlotSpline_PO} This function comapres estimated and emperical selectivity curves for specified Fleet
}
