% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSplines.R
\name{PlotSplines}
\alias{PlotSplines}
\title{Plot size selectivity for SS runs}
\usage{
PlotSplines(
  Path = Path,
  Allfleets,
  f,
  fyears,
  lyears,
  nr,
  nc,
  w,
  h,
  name = "Selex"
)
}
\description{
\code{PlotSplines} This function compares estimated and empirical selectivity curves for specified fleets
}
