% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0_profile.R
\name{R0_profile}
\alias{R0_profile}
\title{Generate R0 likelihoood profile figure}
\usage{
R0_profile(Path, R0, title)
}
\description{
/code{R0_profile} This function makes R0 profile plot for BET SAC11 assessments
}
