% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_diagnostics.R
\name{R_diagnostics}
\alias{R_diagnostics}
\title{Compute R shift for BET (before and after 1994) and mean(R)/R0}
\usage{
R_diagnostics(Rep = Rep)
}
\description{
\code{R_diagnostics} This function computes the ratio of median predicted recruitment before and after 1994; and the ratio of mean recruitment to R0
}
