% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_shift.R
\name{R_shift}
\alias{R_shift}
\title{Compute R shift for BET (before and after 1994)}
\usage{
R_shift(Rep = Rep)
}
\description{
\code{R_shift} This function computes the ratio of median predicted recruitment before and after 1994
}
