% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_compare.R
\name{cpue_compare}
\alias{cpue_compare}
\title{Compare cpue from different VAST models}
\usage{
cpue_compare(
  Path,
  Legend,
  Region,
  Save_Path,
  rescale = TRUE,
  ylabel = "CPUE",
  xlabel = "",
  ylim,
  xlim,
  CV = TRUE,
  w = 15,
  h = 6
)
}
\description{
/code{cpue_compare} This function plot the comaprison of standardized cpue from different VAST models
}
