% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faa.R
\name{faa}
\alias{faa}
\title{Fishing mortality by age group}
\usage{
faa(Dir, Last_Year, Species, Save_Dir)
}
\description{
\code{faa} This function computes the fishing mortality by age group for the stock assessment report
}
