% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_plot.R
\name{impact_plot}
\alias{impact_plot}
\title{Fishing impact plot}
\usage{
impact_plot(Dir, n_year, BaseName, n_fishery, title)
}
\description{
\code{impact_plot} This function automayically generates the fishery impact plot for SS v3.23
}
