% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kobetable_SAC11.R
\name{make_kobetable_SAC11}
\alias{make_kobetable_SAC11}
\title{Make Kobe table for sAC11}
\usage{
make_kobetable_SAC11(fyear, lyear, Path, KobePath, FFleets,
  STD_only = TRUE, newSS, FstdPath, FlimitPath, DynamicPath)
}
\description{
\code{make_kobetable_SAC11} This function automayically generates the table for kobe plot
}
