% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareSelex.R, R/PlotSelex.R
\docType{data}
\name{Path}
\alias{Path}
\title{Compare size selectivity for SS runs}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
Path

Path
}
\description{
/code{CompareSelex} This function plot compare selectivity curves among models

\code{PlotSelex} This function plot estimated selectivity curves for specified fleets
}
\keyword{datasets}
