% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SB.R
\name{plot_SB}
\alias{plot_SB}
\title{Plot quarterly SBR}
\usage{
plot_SB(SS_Dir, lyear, fyear, legend, Save_Dir, ymax, figure_name, title)
}
\description{
\code{plot_SB} This function plots quarterly SBR for the stock assessment report
}
