% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPearsonRes.R, R/R_shift.R
\name{PlotPearsonRes}
\alias{PlotPearsonRes}
\title{Plot Pearson residuals for SS runs}
\usage{
PlotPearsonRes(Rep = Rep)

PlotPearsonRes(Rep = Rep)
}
\description{
\code{PlotPearsonRes} This function makes Pearson residual plots for a specified fishery

\code{R_shift} This function computes the ratio of median predicted recruitment before and after 1994
}
