% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPearsonRes.R
\name{PlotPearsonRes}
\alias{PlotPearsonRes}
\title{Plot Pearson residuals for SS runs}
\usage{
PlotPearsonRes(Rep = Rep, myFleet = myFleet, Path = Path)
}
\description{
\code{PlotPearsonRes} This function makes Pearson residual plots for a specified fishery
}
