% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSelex.R
\name{plotSelex}
\alias{plotSelex}
\title{Plot size selectivity for SS runs}
\usage{
plotSelex(replist, FleetNums, numRows, numCols)
}
\description{
\code{cpue_fit} This function plot estimated selectivity curves for specified fleets
}
