% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSplines.R
\name{PlotSplines}
\alias{PlotSplines}
\title{Plot size selectivity for SS runs}
\usage{
PlotSplines(Rep = Rep, Path = Path, Allfleets = c(1:6, 13:17, 19:20),
  f)
}
\description{
\code{PlotSplines} This function comapres estimated and emperical selectivity curves for specified fleets
}
