% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_data}
\alias{select_data}
\title{Generate area code for longline}
\usage{
select_data(data, c1, c2, c3, c4)
}
\arguments{
\item{c1}{minimal sets by vessel}

\item{c2}{minimal quarters by vessel}

\item{c3}{minimal sets by grid}

\item{c4}{minimal quarters by grid}
}
\description{
\code{data_selection} This function filter the data for LL CPUE standardization
}
